[General] Fix TokenInterceptor.kt and the functions to add car and get user's cars

+++ app/src/main/java/gr/fellow/fellow_traveller/ui/newtrip/NewTripViewModel.kt
@@ -149,7 +149,7 @@ constructor(
                 destinationFrom.value?.placeId.toString(), destinationTo.value?.placeId.toString(),
                 destinationPickUp.value?.placeId.toString(),
                 dateTimeToTimestamp(date.value.toString(), time.value.toString()),
-                pet.value!!, seats.value!!, bags.value!!, message.value.toString(), price.value!!, car.value?.id!!
+                pet.value!!, seats.value!!, bags.value!!, message.value.toString(), price.value!!, car.value?.id!!.toInt()
             )
                 ) {
                 is ResultWrapper.Success -> {
